+++ spec/services/group_draw_remover_spec.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+# rubocop:disable RSpec/NestedGroups
 require 'rails_helper'
 
 RSpec.describe GroupDrawRemover do
@@ -9,15 +10,47 @@ RSpec.describe GroupDrawRemover do
       let(:group) { instance_spy('group', draw_id: 123) }
       let(:membership) { instance_spy('membership') }
       let(:invitation) { instance_spy('membership') }
+      let(:lottery) { instance_spy('lottery_assignment') }
+      let(:clip_membership) { instance_spy('clip_membership') }
 
       before do
         allow(group).to receive(:members).and_return([member])
         allow(group).to receive(:pending_memberships).and_return([invitation])
+        allow(group).to receive(:lottery_assignment).and_return(lottery)
+        allow(group).to receive(:clip_membership).and_return(clip_membership)
       end
+
       it 'removes the draw_id from the group' do
         described_class.remove(group: group)
         expect(group).to have_received(:update!).with(draw_id: nil)
       end
+
+      context 'with lottery assignment' do
+        before { allow(lottery).to receive(:present?).and_return(true) }
+
+        it 'destroys the lottery assignment if not clipped' do
+          allow(lottery).to receive(:groups).and_return([group])
+          described_class.remove(group: group)
+          expect(lottery).to have_received(:destroy!)
+        end
+        it 'nullifies the lottery assignment association if clipped' do
+          other_group = instance_spy('group')
+          allow(lottery).to receive(:groups).and_return([group, other_group])
+          described_class.remove(group: group)
+          expect(group).to have_received(:update!)
+            .with(lottery_assignment_id: nil)
+        end
+      end
+
+      context 'with clip' do
+        before { allow(clip_membership).to receive(:present?).and_return(true) }
+
+        it 'destroys the clip membership' do
+          described_class.remove(group: group)
+          expect(clip_membership).to have_received(:destroy!)
+        end
+      end
+
       it 'removes the draw_id from members and saves old_draw_id' do
         described_class.remove(group: group)
         expect(member).to have_received(:remove_draw)