+++ app/controllers/draws_controller.rb
@@ -117,6 +117,15 @@ class DrawsController < ApplicationController # rubocop:disable ClassLength
     @suites_with_results = SuitesWithRoomsAssignedQuery.new(@draw.suites).call
   end
 
+  def group_export
+    @groups = @draw.groups.includes(:lottery_assignment)
+                   .order('lottery_assignments.number')
+    attributes = %i(name lottery_number suite_number)
+    result = CSVGenerator.generate(data: @groups, attributes: attributes,
+                                   name: 'groups')
+    handle_file_action(**result)
+  end
+
   private
 
   def authorize!